Remove unused createCreditCardMethod
