+++ packages/composite-checkout/src/public-api.js
@@ -27,7 +27,6 @@ import { createFreePaymentMethod } from './lib/payment-methods/free-purchase';
 import { createStripeMethod } from './lib/payment-methods/stripe-credit-card-fields';
 import { createApplePayMethod } from './lib/payment-methods/apple-pay';
 import { createPayPalMethod } from './lib/payment-methods/paypal';
-import { createCreditCardMethod } from './lib/payment-methods/credit-card';
 import { createExistingCardMethod } from './lib/payment-methods/existing-credit-card';
 import { useActiveStep, useIsStepActive, useIsStepComplete } from './lib/active-step';
 import CheckoutOrderSummary, {
@@ -53,7 +52,6 @@ export {
 	OrderReviewSection,
 	OrderReviewTotal,
 	createApplePayMethod,
-	createCreditCardMethod,
 	createExistingCardMethod,
 	createFreePaymentMethod,
 	createFullCreditsMethod,
